+++ mex/editor/client/src/app/shared/to-label-pipe.ts
@@ -1,11 +1,12 @@
 import type { PipeTransform } from "@angular/core";
 import { Pipe } from "@angular/core";
 import type { Text } from "./models/shared";
-import type { BilingualText, Concept } from "./models";
-import { type CreateItem, type PreviewItem } from "./models";
+import type { BilingualText, Concept } from "./models/concept";
 import type { PreviewOrganizationalUnit } from "./models/organizational-unit";
 import type { PreviewPerson } from "./models/person";
 import type { PreviewContactPoint } from "./models/contact-point";
+import type { PreviewItem } from "./models";
+import type { CreateItem } from "./models/create-item";
 
 @Pipe({
   name: "toLabel",
